Add files via upload
Jupyter notebooks to generate Maps using GeoPandas tool. 

Map_Nacimientos Notebook: The purpose of this code is to generate a map that reflects the percentage of young mothers using the database provided by the Ministry of Health at the provincial level.

Map: The purpose of this code is to generate a map that captures early fertility rates at the provincial level using the ENDES databases.